Refactor aggregation strategies into dedicated module
